passport_api/requirements.txt
Django==3.2.6
psycopg2-binary==2.8.6
gunicorn==20.1.0
djangorestframework==3.12.4
drf-extra-fields==3.1.1
pillow==8.3.2
djoser==2.1.0
djangorestframework_simplejwt==4.8.0
python-dotenv==0.19.1
drf-yasg==1.20.0